sim/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "port26-sim"
version = "0.1.0"
description = "Downscaling pipeline for coarse meteorology and dense station observations"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cdsapi>=0.7.5",
    "cfgrib>=0.9.15.1",
    "contextily>=1.6.2",
    "eccodes>=2.46.0",
    "ecmwf-opendata>=0.3.26",
    "fmiopendata>=0.5.0",
    "joblib>=1.4.2",
    "manim>=0.20.1",
    "matplotlib>=3.10.0",
    "netCDF4>=1.7.4",
    "numpy>=2.2.2",
    "pandas>=2.2.3",
    "planetary-computer>=1.0.0",
    "pyarrow>=18.1.0",
    "pystac-client>=0.8.5",
    "rasterio>=1.4.3",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.1",
    "xarray>=2025.1.2",
    "xgboost>=2.1.4",
]

[tool.setuptools]
include-package-data = true
packages = [
    "sim",
    "sim.data_in",
    "sim.inference",
    "sim.models",
    "sim.preprocess",
    "sim.workflows",
]

[tool.setuptools.package-dir]
sim = "."
